condominium, address, worker, authorization level CRUD

+++ src/main/java/com/jccv/tuprivadaapp/repository/TokenRepository.java
@@ -7,13 +7,13 @@ import org.springframework.data.jpa.repository.Query;
 import java.util.List;
 import java.util.Optional;
 
-public interface TokenRepository extends JpaRepository<Token, Integer> {
+public interface TokenRepository extends JpaRepository<Token, Long> {
 
     @Query("""
         select t from Token t inner join User u on t.user.id = u.id
         where t.user.id = :userId and t.loggedOut = false
         """)
-    List<Token> findAllTokensByUser(Integer userId);
+    List<Token> findAllTokensByUser(Long userId);
 
     Optional<Token> findByToken(String token);
 }
